requirements.txt
setuptools
pywebview~=5.2
pyinstaller; sys_platform != 'darwin'
py2app; sys_platform == 'darwin'
pandas~=2.2.3